requirements.txt
# Dependencias principales del Framework Django
Django==5.1.2                    # Framework web principal
django-crispy-forms==2.1         # Para mejorar el renderizado de formularios
crispy-bootstrap5==2024.2        # Integración de Crispy Forms con Bootstrap 5
djangorestframework==3.14.0      # Framework para crear APIs REST

# Framework FastAPI y servidor ASGI
fastapi==0.109.1                 # Framework alternativo para APIs
uvicorn==0.27.0                 # Servidor ASGI para FastAPI

# Base de datos y autenticación
psycopg2-binary==2.9.9          # Adaptador PostgreSQL para Python
python-jose[cryptography]==3.3.0 # Para manejo de JWT (JSON Web Tokens)
passlib[bcrypt]==1.7.4          # Para hash de contraseñas
python-multipart==0.0.6         # Para manejo de formularios multipart

# Utilidades y herramientas
python-dotenv==1.0.0            # Para manejo de variables de entorno
django-cors-headers==4.3.1       # Para manejo de CORS en Django
Pillow==10.2.0                  # Para procesamiento de imágenes
django-allauth==0.61.0          # Para autenticación social y manejo de cuentas